refactor: #686 analysis_scoring_service を集計境界で分割する
スコア計算・レコメンド文言・埋め込み数学をファイル分割し、
計算式は変更しない。公開ラッパと既存テストは維持する。
